refactor: ui/component 추가 및 ValidateTextField 이동
