+++ app/src/main/java/nextstep/signup/ui/component/ValidatedTextField.kt
@@ -1,0 +1,46 @@
+package nextstep.signup.ui.component
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextField
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun ValidatedTextField(
+    field: InputFieldModel,
+    modifier: Modifier = Modifier,
+) {
+    var error by remember { mutableStateOf<String?>(null) }
+    Column(modifier = modifier) {
+        TextField(
+            modifier = Modifier.fillMaxWidth(),
+            value = field.value,
+            label = field.label,
+            isError = error != null,
+            onValueChange = {
+                field.onValueChange(it)
+                val result = field.validator.validate(it)
+                error = if (!result.isValid) result.message else null
+            },
+            singleLine = true,
+            visualTransformation = field.visualTransformation
+        )
+        error?.let {
+            Text(
+                text = it,
+                color = MaterialTheme.colorScheme.error,
+                style = MaterialTheme.typography.labelSmall,
+                modifier = Modifier.padding(start = 16.dp, top = 4.dp)
+            )
+        }
+    }
+}
